+++ VComponents/Components/Core/Base Header Footer/VBaseHeaderFooterView.swift
@@ -61,9 +61,9 @@ public struct VBaseHeaderFooter: View { // FIXME: Remove
     }
 }
 
-// MARK: - Preview
-struct VBaseHeaderFooter_Previews: PreviewProvider {
-    static var previews: some View {
-        VSectionList_Previews.previews
-    }
-}
+//// MARK: - Preview
+//struct VBaseHeaderFooter_Previews: PreviewProvider {
+//    static var previews: some View {
+//        VSectionList_Previews.previews
+//    }
+//}