+++ VComponents/Components/Lists/Base List/IdentifiableElement.swift
@@ -1,5 +1,5 @@
 //
-//  VBaseListElement.swift
+//  IdentifiableElement.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 1/10/21.
@@ -7,8 +7,8 @@
 
 import Foundation
 
-// MARK: - V Base List Element
-struct VBaseListElement<ID, Value>: Identifiable where ID: Hashable {
+// MARK: - Identifiable Element
+struct IdentifiableElement<ID, Value>: Identifiable where ID: Hashable {
     // MARK: Properties
     let id: ID
     let value: Value