+++ VComponentsDemo/Scenes/Home/HomeView.swift
@@ -23,7 +23,7 @@ struct HomeView: View {
         .init(id: 2, title: "Item Pickers", rows: [.segmentedPicker, .menuPicker, .wheelPicker]),
         .init(id: 3, title: "Value Pickers", rows: [.stepper, .slider, .rangeSlider]),
         .init(id: 4, title: "Inputs", rows: [.textField]),
-        .init(id: 5, title: "Lists", rows: [.baseList, .list, .sectionList, .accordion]),
+        .init(id: 5, title: "Lists", rows: [.baseList, .accordion]),
         .init(id: 6, title: "Modals", rows: [.modal, .halfModal, .sideBar, .dialog, .menu, .actionSheet]),
         .init(id: 7, title: "Messages", rows: [.toast]),
         .init(id: 8, title: "Indicators", rows: [.spinner, .progressBar, .pageIndicator]),
@@ -36,7 +36,7 @@ struct HomeView: View {
         case segmentedPicker, menuPicker, wheelPicker
         case stepper, slider,  rangeSlider
         case textField
-        case baseList, list, sectionList, accordion
+        case baseList, accordion
         case modal, halfModal, sideBar, dialog, menu, actionSheet
         case toast
         case spinner, progressBar, pageIndicator
@@ -67,8 +67,6 @@ struct HomeView: View {
             case .textField: return VTextFieldDemoView.navBarTitle
                 
             case .baseList: return VBaseListDemoView.navBarTitle
-            case .list: return VListDemoView.navBarTitle
-            case .sectionList: return VSectionListDemoView.navBarTitle
             case .accordion: return VAccordionDemoView.navBarTitle
 
             case .modal: return VModalDemoView.navBarTitle
@@ -115,8 +113,6 @@ struct HomeView: View {
             case .textField: VTextFieldDemoView()
 
             case .baseList: VBaseListDemoView()
-            case .list: VListDemoView()
-            case .sectionList: VSectionListDemoView()
             case .accordion: VAccordionDemoView()
 
             case .modal: VModalDemoView()